cleanup of the added settings code
