+++ HumanSpaceProgram/Assets/HSP.Settings/GameSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2fd225d7b57b6604e9fc4a03bd36b021
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
