+++ HumanSpaceProgram/Assets/HSP.Timelines/ScenarioSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90ee17f8e5401214e85c9c1edce2dbd9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
