--- HumanSpaceProgram/Assets/HSP.Timelines/Scenarios.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2e0633e6dc1ccc84e8df3162595d0258
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
