--- HumanSpaceProgram/Assets/HSP.Timelines/Scenarios/Scenario.cs
@@ -1,13 +1,0 @@
-﻿using HSP.Timelines.Serialization;
-using System;
-using System.Collections.Generic;
-
-namespace HSP.Timelines.Scenarios
-{
-    public class Scenario
-    {
-        ScenarioMetadata _metadata;
-
-        List<Type> _shownScenarioSettings;
-    }
-}
