+++ HumanSpaceProgram/Assets/HSP.Timelines/TimelineSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 94dd2492664f44043938119f3866f41c
+guid: 562e77200595a254d9d4877715fc61c9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
