+++ HumanSpaceProgram/Assets/HSP.Timelines/___HSP.Timelines.asmdef
@@ -9,7 +9,8 @@
         "GUID:f00983dade8147042b611005c376fd7e",
         "GUID:2a88aeff819575d4eb14afd4a2966bb9",
         "GUID:53cb8ceb0f9a3d94ab671037324752c6",
-        "GUID:2f7dd5a314d78174bab3f10d91029679"
+        "GUID:2f7dd5a314d78174bab3f10d91029679",
+        "GUID:8cd1ef44bbf901747bea33ff82069ef5"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
